1. I changed the file name for EEPROM utility to avoid the name conflict with the standard header. 2. In order to use the correct library, I explicitly specified the dependent standard library in the platformio.ini file.
