+++ Grbl_Esp32/eeprom_memcpy.cpp
@@ -1,5 +1,5 @@
 /*
-  eeprom.h - Header for system level commands and real-time processes
+  eeprom_memcpy.cpp - Header for system level commands and real-time processes
   Part of Grbl
   Copyright (c) 2014-2016 Sungeun K. Jeon for Gnea Research LLC
 	
@@ -18,14 +18,26 @@
   along with Grbl.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#ifndef eeprom_h
-#define eeprom_h
-
 #include "grbl.h"
 
-//unsigned char eeprom_get_char(unsigned int addr);
-//void eeprom_put_char(unsigned int addr, unsigned char new_value);
-void memcpy_to_eeprom_with_checksum(unsigned int destination, char *source, unsigned int size);
-int memcpy_from_eeprom_with_checksum(char *destination, unsigned int source, unsigned int size);
+void memcpy_to_eeprom_with_checksum(unsigned int destination, char *source, unsigned int size) {
+  unsigned char checksum = 0;
+  for(; size > 0; size--) { 
+    checksum = (checksum << 1) || (checksum >> 7);
+    checksum += *source;
+    EEPROM.write(destination++, *(source++)); 
+  }
+  EEPROM.write(destination, checksum);
+  EEPROM.commit();
+}
 
-#endif
+int memcpy_from_eeprom_with_checksum(char *destination, unsigned int source, unsigned int size) {
+  unsigned char data, checksum = 0;
+  for(; size > 0; size--) { 
+    data = EEPROM.read(source++);
+    checksum = (checksum << 1) || (checksum >> 7);
+    checksum += data;    
+    *(destination++) = data; 
+  }
+  return(checksum == EEPROM.read(source));
+}
