+++ platformio.ini
@@ -3,9 +3,21 @@ src_dir=Grbl_Esp32
 lib_dir=libraries
 data_dir=Grbl_Esp32/data
 
+[common_env_data]
+lib_deps_builtin =
+	EEPROM
+	BluetoothSerial
+	WiFi
+	FS
+	SD
+	SPI
+	Preferences
+	SPIFFS
+	WebServer
+
 [env:nodemcu-32s]
 platform = espressif32
 board = nodemcu-32s
 framework = arduino
 upload_speed = 512000
-
+board_build.partitions = no_ota.csv
